LPD-90478 less code
